pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = "Uses Neo4j's graphrag package to create a knowledge graph from an Obsidian vault."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.28.2",
    "neo4j-graphrag>=1.9.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "watchdog>=3.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "watchdog>=6.0.0",
]
